Player, geração de canos, colisão e game over

+++ Flappy-bird/src/com/gcstudios/graficos/UI.java
@@ -12,7 +12,7 @@ public class UI {
 	public void render(Graphics g) {
 		g.setColor(Color.white);
 		g.setFont(new Font("arial",Font.BOLD,18));
-		g.drawString("Score: " + Game.score,20,20);
+		g.drawString("Score: " + (int)Game.score,20,20);
 	}
 	
 }
